requirements.txt
pyspark == 3.4.1
setuptools
numpy
pandas
shap
matplotlib
scikit-learn
synapseml
```